requirements.txt
pymongo
pyrogram
TgCrypto
python-dotenv